+++ rules/camel-case-dash.js
@@ -1,1 +1,1 @@
-module.exports = require('../helpers').createCamelCaseRule('-')
+module.exports = require('../helpers/createCamelCaseRule')('-')
